Fix exopackage test's submap call
Summary:
The intention here is to get a submap containing everything in a
directory by constructing a string that will sort after any actual
contents of the directory. This code is trying to add the character for
code 255 to the directory, but `dirPath +
Character.toChars(255).toString()` is actually going to be something
like `dirPath + "[C@1540e19d"`. That happens to work, but is kind of
silly. Change it to actually do what it was intending to do.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 1e2ca43