Remove AbstractSourcePath and make ArchiveMemberSourcePath immutable
Summary:
Now that asReference is removed, removes compareReferences. This was the last remaining use of AbstractSourcePath.
This makes the Pair in ArchiveMemberSourcePath unnecessary, so we can separate the fields and make it immutable.

Test Plan:
Run 'Debug Buck' locally
The only test changes were to change compareReferences to compareTo in PathSourcePathTest

Reviewed By: asp2insp, illicitonion

fbshipit-source-id: 9921cb9