Add ResourceTable.slice()
Summary:
This allows extracting a part of a ResourceTable into another. Given a
map of type->type_count, it will create a ResourceTable consisting of
all the same types with the first type_count resources extracted. Any
configurations that no longer have a resource with a value are discarded.

The stringpools are currently unchanged (a followup will make it so that
we determine which strings are used and extract just those strings and
rewrite the references to them).

Test Plan: Added tests. CI

Reviewed By: dreiss

fbshipit-source-id: 69c8152