Make ResourceTable.slice optimize strings
Summary:
This removes strings from the main and keys stringpools that are no
longer used. It doesn't optimize the types stringpool, but that one is
tiny and didn't seem necessary to optimize.

Determines the used strings by just iterating over the references in the
resource table.

Test Plan: Added tests. CI

Reviewed By: dreiss

fbshipit-source-id: ec82e11