Do source-ABI compatibility filtering earlier
Summary:
In our attempts to make a class-based ABI binary identical to the source-based ones,
we had `SourceAbiCompatibleVisitor` at the very end of our visitor chain when generating
class-based ABIs, suppressing any events that would not have been present from source-based
ones. However, some such events (bridge methods) result in the ABI changing in other ways
(InnerClass references). Thus, it's more correct to do the filtering at the beginning of
the visitor chain, so that all the rest of the ABI generation is working off the same inputs
regardless of whether the ABI is source or class based.

Test Plan: New unit test

Reviewed By: dreiss

fbshipit-source-id: 6ef6d37