Sort InnerClasses table
Summary:
Class-based ABIs are created by stripping down the full classes. If we don't
do something special, then, the order of the `InnerClasses` table would be affected
by references to inner classes inside method bodies. Replicating that in source-based
ABIs would be a pain, so we just sort the table.

We have to be careful about the sort order, though, because at least for member
classes it determines the iteration order of `Element.getEnclosedElements`.

Test Plan: Updated unit test

Reviewed By: dreiss

fbshipit-source-id: 9b3d365