Enable source ABI generation (with dependencies)
Summary:
Source-based ABI jars are now binary identical to class-based ones
(with a couple harmless exceptions; see `SourceAbiCompatibleVisitor`).
Turning them on for Buck itself.

As of now this is at best performance neutral. As we do performance
work on the source ABI generation itself and move on to ABI generation
without deps we should see performance improvements.

Test Plan: CI

Reviewed By: Coneko, dreiss

fbshipit-source-id: e66d506