add cache mode to cacheResult
Summary:
This adds ArtifactCacheMode (http, dir etc) into the CacheResult.
This allows to know what cache returned the CacheResult and it
will be used for logging and debugging purposes in future diffs.

This also removes the generation of a RuntimeException in the
CachingBuildEngine in the case that the cache throws an exception.
Afaik we never hit that case and most probably the code was from a
different era where that was a possibility.

In order to avoid some major `artifact_cache` BUCK file refactoring I
separated the enum to a different class.

Test Plan: - tweaked the tests

fbshipit-source-id: b88c111