Add comments and change some names
Summary:
The naming of these fields (derived from
cxx.cflags/cppflags/cxxflags/cxxppflags) can be really unintuitive.
This just adds a comment about the naming and what the different flags
are used for and then renames some functions to make it more clear what
the flags are for.

Test Plan: Just naming+comments

Reviewed By: mzlee

fbshipit-source-id: 552b0af