Return clean errors on unknown identifier
Summary:
The source-generated ABI code throws exceptions when `ErrorType`s are encountered, which
means it would fail the step with a useless exception when trying to generate ABIs for code
that had unknown identifiers in it. In practice this didn't show up much, as the full
compilation is happening at the same time and Buck seems to prefer displaying the errors
from that for some reason.

Test Plan:
Add a public field of a bogus type to some class
Compile it with `-j 1` to give the source ABI the best chance of running first (might have to try
a couple times)
Observe useful compiler error instead of useless exception

Reviewed By: dreiss

fbshipit-source-id: 05ed368