Don't use Trees.getPath
Summary:
Grabbing some performance low-hanging fruit.

`Trees.getPath` does a depth-first search of the AST of the enclosing file in
order to construct the path, and it was being called on every `Element` during source ABI
construction.

`Trees.getTree` is still not optimal (it iterates all members of the enclosing class), but
it's an easy switch and it's at least not searching the entire AST.

If this isn't fast enough, longer-term we might look into keeping track of the trees as we
enter them. There's some gotchas there that we've run into previously, for which I've added
a comment.

Test Plan:
CI for regression detection
YourKit profile of clean `buck build buck` before and after to verify performance improvement

Reviewed By: dreiss

fbshipit-source-id: 50a3bb6