Pass in environment via Autosparse to Mercurial
Summary:
Some mercurial commands need to authenticate to the backend, and that relies on
environment variables we currently don't pass in via the Autosparse route.

This isn't a problem if chg (long-running Mercurial process) is already
running, but if it isn't then auth can fail. Pass in the client environment
wherever possible.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 70d3e70