Chrome Trace: Fix coloring of rule_key_calc
Summary:
`rule_key_calc` events were showing up in different colors in the trace. The color
of a trace slice of a given name is supposed to be constant across and between traces.

The root cause was that in an effort to be efficient with events, `BuildRuleEvent.FinishedRuleKeyCalc`
implements both `BuildRuleEvent.Suspended` and `RuleKeyCalculationEvent.Finished`.
`EventBus`'s dispatch logic dispatches from most to least general, so the handler
for `BuildRuleEvent.Suspended` was being invoked before the one for
`RuleKeyCalculationEvent.Finished`. This resulted in an end event being written for
the build rule slice *before* the end event for the rule key calc slice, which confused
`trace-viewer`. Fortunately, the impact of this was just incorrect coloring of the
slices.

This is a hacky workaround because at the present moment I don't want to bite off the
more general problem of being efficient with events while retaining nesting order,
or whether this optimization is necessary in the first place.

Test Plan: Do a build, observe `rule_key_calc` slices are now all the same color.

Reviewed By: asp2insp

fbshipit-source-id: fd2be05