rust: dedup command line args generated for dependencies
Summary:
It's very common for the same command-line to be passed repeatedly, redundantly - for example the library path if all the libraries are in the same directory. Accumulate the Args from dependencies into a LinkedHashSet so that their relative order is preserved, but
duplicates are removed.

Note that this requires each logical argument to be one actual element on the command line (ie, "-Lfoo", not ["-L", "foo"]), so make sure that's true in the various places args are generated.

Dedup is not applied to user-supplied arguments, so there's no concern about splitting them up.

Test Plan: Unit tests

Reviewed By: marcinkosiba

fbshipit-source-id: 31215f4