Factor out MergedNativeLibraryConstituents.isActuallyMerged
Summary:
We construct a MergedNativeLibraryConstituents object for every native
library that will end up in the apk, and we use the presence of an
explicit soname to distinguish a library defined by the merge config
(even if it only has one constituent) from one that's a normal dep of
the app.  Create a helper method to make it more obvious that this is
happening.

Test Plan: CI

fbshipit-source-id: 7fa2ac3