Fix ndk.extra_***flags.
Summary:
This makes ndk.extra_***flags actually be added to the flags last so
that they can override the defaults. It reorders the flags a bit to be more
logical (flags should be added in approximately most general to most specific
order).

This enables a bunch of checks in NdkCxxPlatformsTest that were previously failing.

Test Plan: CI + careful review of changes to ndkcxxplatforms.flags test file

Reviewed By: marcinkosiba

fbshipit-source-id: 0ca6fc7