Factor out short-circuiting logic
Summary:
This introduces CachingBuildEngine.transformBuildResultIfNotPresent() and
transformBuildResultAsyncIfNotPresent(). These handle the
short-circuiting if we already have a BuildResult or if we've already
seen a failure. Also, the less async version handles wrapping in a
resume-suspend scope.

This actually has a couple of side effects:
1. rulekey cache checks (on the cacheActivityService) are wrapped in
resume-suspend scopes and so properly show up in SuperConsole and
traces.
2. short-circuiting based on firstFailure is now more consistent and may
happen at an earlier stage than previously.

Test Plan:
Manually verify that rulekey fetches now show in superconsole and have the rule marked in chrome traces.
CI

Reviewed By: marcinkosiba

fbshipit-source-id: 42edbc8