Shorten ProjectFilesystem.toString
Summary:
It's rarely useful to see the full list of blacklisted paths, so
eliminate it from ProjectFilesystem.toString, but still include a hash,
because they are used in .equals(), and it would be nice to avoid having
to non-equal PFS instances with the same string representation.

This makes my logs easier to read.

Test Plan: Looked at logs.

Reviewed By: Coneko

fbshipit-source-id: e77f032