Don't include synthetic classes
Summary:
`javac` 8 generates its synthetic classes with null `outerName` and `innerName`, meaning
that we skip them because we think they're anonymous or local classes. `javac` 7 generates
some of them with an `outerName`, so we don't.

Test Plan:
Build a class that instantiates a private static member class, resulting in a synthetic
class being emitted, using `javac` 7, with ABI verification on. Observe verification passes.

Reviewed By: dreiss

fbshipit-source-id: fec186c