Fix ndk.extra_***flags (take 2).
Summary:
This makes ndk.extra_***flags actually be added to the flags last so
that they can override the defaults.

This does not reorder the flags to what I think is a more logical order, because there's not much value and I don't want to chase down the things that that might break.

This enables a bunch of checks in NdkCxxPlatformsTest that were previously failing.

Test Plan: CI + careful review of changes to ndkcxxplatforms.flags test file

Reviewed By: mzlee

fbshipit-source-id: d1bbc2e