JarBuilder: Don't leak ZipFileSystems when the full jar contains resources
Summary:
`JarBuilder` holds on to `InputStream` suppliers so that it can write its
entries in sorted order. When the full source was coming from a jar, those
suppliers were calling methods on `JarReader`. The actual writing of
the entries occurs in `FilesystemStubJarWriter.close`, which is called
after the `JarReader` and its associated file system have been closed.
However, due to `JarReader`s implementation, it would just happily reopen (and
never close) the filesystem.

This was discovered through an intermittent integration test failure on
Windows (where Buck was failing to delete an output directory due to the
jar file still being open).

This diff adds a check in `JarReader` so that this error would be a crash
on all platforms, and changes `StubJar` to hold the `JarReader` open
until after the `StubJarWriter` has been closed.

Test Plan: `GenruleDepsIntegrationTest`

Reviewed By: dreiss

fbshipit-source-id: e718222