Add more info for cannot delete directory failures
Summary:
This form of failure often happens intermittently on Windows
because the file system will not allow files to be deleted
while a process has them open. Typically there's a file
stream or similar that has not been closed. Listing the
files in the exception message can help direct the investigation,
especially when the tests are being run in CI and developers
don't have their own Windows machines.

Test Plan: Used this to help track down one such failure

Reviewed By: asp2insp

fbshipit-source-id: 800b939