ClassLoaderBootstrapper: Create ClassLoader earlier
Summary:
When an agent such as YourKit is present in the process, it may cause `java.util.logging`
to be initialized earlier than usual, even before `main` is run. That results in a
`NullPointerException` as `com.facebook.buck.cli.bootstrapper.LogConfig` attempts to
load `com.facebook.buck.log.LogConfig`. That exception is swallowed, and simply results
in no logging.

Test Plan:
Run Buck with the YourKit agent attached from the start, observe `buck-0.log` and friends
are created and have content.

Reviewed By: dreiss

fbshipit-source-id: 013df61