Close cache at the end of cache command
Summary:
Currently, when the buck cache command is running, the artifact cache are
created but never close.
This diff forces it to be closed.
It will trigger the cache clean up properly at the end of the cache command.

Test Plan:
Change DirArtifactCache.close() to always call DirArtifactCache.deleteOldFiles().

```
$ buck cache 1234
```

Verified that DirArtifactCache.deleteOldFiles() is called.

Reviewed By: yiding

fbshipit-source-id: 7968aca