Fix crash building Android rules
Summary:
The `DefaultJavaLibraryBuilder` needs all of the dependencies of the original
rule in order to generate the source ABI rule properly (for now). That includes the
`DummyRDotJava` rule. However, the code for getting the `DummyRDotJava` rule name
wasn't properly handling the case of needing to get the `DummyR` corresponding to a
`#source-abi` rule.

Test Plan: Built one of our Android apps

Reviewed By: dreiss

fbshipit-source-id: 435f17a