WorkerProcessIdentity for persistence purposes
Summary:
Previously we had 2 optional fields on WorkerProcessParams - key and hash. But we required them both to be either present or not present.
To make code a bit cleaner, we combine them into a single object which is now optional. Thus, we eliminate a possible unexpected or ambiguous behaviour when we set key and omit hash or vice versa.

Test Plan: CI

Reviewed By: illicitonion

fbshipit-source-id: f34614f