Don't exclude the package-info class
Summary:
When annotations are present on a package, `javac` generates a
synthetic `package-info` class which we were stripping out incorrectly.

This commit stops stripping it from class ABIs. A separate commit
will start generating the class for source ABIs; I wanted to get this
one in quickly so that we have at least one from of ABI generation that
is correct.

Test Plan: New unit test

Reviewed By: asp2insp

fbshipit-source-id: b7d6f95