Workaround SuperConsole bug
Summary:
The SuperConsole, when a user has configured ui.thread_line_limit to a
very large value, doesn't actually work very well when we tell it about
a lot of threads.

This is a workaround so that these cache threads will not be reported to
SuperConsole (or, sadly, to Chrome traces).

Test Plan: CI, manually run `buck build -c ui.thread_line_limit=1000 fb4a` and see no overflowing

Reviewed By: marcinkosiba

fbshipit-source-id: 5fc3b1a