ensure runtime deps for AndroidInstrumentationTest include the original AndroidBinary rule
Summary:
currently, the AndroidInstrumentationTest just adds the runtime deps for the input to the rule and one level up for the case of an apk_genrule.
However, if it is a chain of apk_genrules, then we need to keep going until we hit an AndroidBinary rule where the AndroidManifest and ExoPackage files are built.

Test Plan: testing with a revert of D5107242, which exposes this issue

Reviewed By: newobj

fbshipit-source-id: 5e9e40e