PostEnterTaskListener should provide top-level Elements rather than TypeElements
Summary:
Previously the API assumed that the only interesting thing coming out of Enter would be `TypeElement`s.
This is incorrect -- if a `package-info.java` class is present, a `PackageElement` will also be there.
This commit changes the API signature to use `Element` and propagates that change up the call graph. No
logic changes yet.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 737dc79