VersionTargetGraphBuilderTest should construct target graphs exactly as specified
Summary:
TargetGraphFactory by default constructs a flavorless node in addition to the
flavored node that is passed in. I add a method to construct a target node only
of what is passed in. I think that is in line with what this test is testing.

I am fixing the semantics of TargetGraphFactory.newInstance in a future diff,
and this preemptivly fixes this test which breaks with that change.

Test Plan: CI

Reviewed By: beefon

fbshipit-source-id: 71a5fc4