Fix TargetGraphFactory generating invalid target graphs for tests
Summary:
TargetGraphFactory.newInstance generates a graph that contains, in
addition to the nodes passed in, nodes which are unflavored. However the
old logic doesn't properly transform the TargetNode itself to be
unflavored, which seems like incorrect behavior and causes problem for
me in a future diff. This diff corrects that behavior.

ProjectGeneratorTest is changed as the old code will call requireRule on
all nodes, which now will do it to the true flavorless nodes, and fail.
I change this code to be more inline with how it's actually used, where
the target node's build rule is generated as its requested.

Test Plan: CI

Reviewed By: beefon

fbshipit-source-id: 91d7c25