In AbstractCxxSymlinkTreeHeaders, use Stream.sorted() instead of creating an ImmutableSortedSet
Summary:
This code used to create an `ImmutableSortedSet` just to get a stable
ordering for iteration. This is expensive. I change this to use `Stream.sorted`,
which is much faster as it internally uses a simple `Arrays.sort`.

Test Plan: ran a build while profiling, see that this method's time taken is dramatically reduced.

Reviewed By: beefon

fbshipit-source-id: 8a5543f