Stop passing swiftCompanionBuildRule dependency over extraCxxDeps
Summary:
This is already passed within the delegate arg object, passing it via
extraCxxDeps is redundant.

The entry can also be added to the builder directly, instead of also cloning
the original arg object with the change.

Test Plan:
SwiftIOSBundleIntegrationTest still pass. This test fails if the
dependency is removed entirely.

Reviewed By: beefon

fbshipit-source-id: f979e46