Make field counting in DalvikAwareOutputStreamHelper precise
Summary:
Previously, I was missing two sources of field references: enums in
annotations and the multianewarray opcode (which DX converts into a call
to Array.getInstance with an argument like Integer.TYPE).  This tracks
both of them, and updates the field limit to 64k.

Most of the code is just to cover all the cases where annotations can
appear.

Test Plan:
Ran it on all of the jars in FB4A and compared the output to the field
references listed by baksmali.

Reviewed By: sdwilsh

fbshipit-source-id: 609fc22