Synchronize file materialization on per-file locks.
Summary:
We used to materialize all regular files without any synchronization because we
do not want to download each file from the CAS one-by-one. One un-noticed caveat
was that when we get 2 concurrent requests for the same file and we materialize
it twice, the file appears incomplete for a small window of time while we're
re-writing the file using a BufferedOutputStream. This problem got amplified
when we added the global source file dir-cache, because now we immediately copy
the file over to the dir-cache and this timing tends to align well with our
vulnerable window, and we stored a bad copy of the file in dir-cache, and use it
for all consecutive builds.

This diff synchronizes file materialization on a per-file basis.

Test Plan:
Without this diff, global dir cache tends to have mutiple zero byte
files. Maybe this is not the full fix, but it is at least one of the factors
involved in this bug.

Reviewed By: alisdair04

fbshipit-source-id: 286a97f