CachingBuildEngine now logs all build rule failures as they happen
Summary:
==================
Existing behavior
==================

Ignoring ad-hoc logging performed by Steps, all logging of build failures:

- only happens at the end of the build, not when the failures occurs
- only includes details about the first failure that occured (for one or all of the targets given on the command line)
- potentially includes no error message

Specifically, the logging of failures depends on the combination of --keepGoing and --verbose flags:

------------------------------------------------------

-- keepGoing=False
-- verbose=any

At the end of the build, for the first target given on the command line that failed, publish a Console.SEVERE event with the name of the target, along with the first build rule exception that caused the target to fail. Also log exception at DEBUG level.

All other build rule failures are not logged, except on an ad-hoc basis without associated rule name.

------------------------------------------------------

-- keepGoing=True
-- verbose=default

At the end of the build, for every target given on the command line that failed, publish a Console.INFO event with the name of the target.

All other build rule failures are not logged, except on an ad-hoc basis without associated rule name.

------------------------------------------------------

-- keepGoing=True
-- verbose=4 (or above)

At the end of the build, for every target given on the command line that failed, publish a Console.INFO event with the name of the target, along with the first build rule exception that caused the target to fail.

All other build rule failures are not logged, except on an ad-hoc basis without associated rule name.

==================
New behavior
==================

When a build rule failure occurs, the failing rule along with its error text is immediately logged with DEBUG log level.

Additionally, if the console_log_build_rule_failures_inline option in the [build] section of .buckconfig is set to true (default: false), the same information will be published to the console with level SEVERE.

Note: this diff does not change the existing behavior for logging errors at the end of the builds as described above, so at the end of the build, depending on keepGoing/verbose settings, there may be some duplication (but only for the single rule that failed first).

Test Plan:
buck build <some broken target>
- ensure that all errors are logged to the log file and console as they happen, regardless of keepGoing or verbosity mode.

Logging should have the following format and be log.DEBUG/Console.SEVERE level;

  Building rule [<build rule>] failed. Caused by [<exception class>]:
  <Error text, e.g. stderr output of shell script>

fbshipit-source-id: f7e11c6