Hack to resolve the issue of diverging transitive CxxPreprocessorInput implementations
Summary:
getTransitiveCxxPreprocessorInput either gets it from CxxLibrary, if
that rule exists at the point of the call, or computes manually. This
is necessary since AppleTest don't have a CxxLibrary rule. This existing
hack is problematic as the code to generate this result must be
identical between CxxLibrary and the inlined code.

Here I change the hack to be a little bit safer. By passing in the
method to compute this value, I can let apple_library and cxx_library will both
always use CxxLibrary, while apple_test always use the inlined version.

Test Plan: CI

Reviewed By: beefon

fbshipit-source-id: 61c7cd2