Add integration test for using `cxx_precompiled_header` in `apple_library`
Summary: I was getting a stack trace when using prefix headers in apple libraries, so I wanted to make sure that the feature worked as I thought it did. The best way to do that - integration tests. This diff just adds a simple integration test to verify the issue. All it does is use a `cxx_precompiled_header` as the parameter for `apple_library`'s `prefix_header`.

Test Plan: Integration tests

Reviewed By: elsteveogrande

fbshipit-source-id: 2c60b73