Add MoreAsserts.assertStepsNames
Summary:
This is just a slightly better version of asserting the size of a list
of steps. In either case you probably just have to update the expected
value when you make seemingly unrelated changes, but at least in this
case you can see how the list changed (and verify it's as expected).
Also, you can clearly see that grabbing a particular step from the list
is getting the one you expect.

Updated GenruleTest/ApkGenruleTest to use this instead of asserting on
the size of steps.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: fc470e1