Add Precomputed Work to the Macro Expanders
Summary:
This diff introduces a new type of macro expander that is typed and can
precompute expensive work. This precomputed work must be cached by the caller since
macro expanders are stateless. This will be used further up the stack by the query macros
so that they can cache the query results, and avoid the cost of re-evaluating four times.

Test Plan: updated unit tests

Reviewed By: styurin

fbshipit-source-id: c07a309