Convert QueryMacroExpanders to Precomputed
Summary:
This diff allows the query macro expanders to precompute their query results. This
is important because the query is otherwise evaluated four times for MISS (once during action graph, once each for rulekey and input based rule key calculations, and once for the build)
With precomputation, the query can be evaluated once.

Test Plan: updated tests to ensure precomputed work is cached and not recomputed.

Reviewed By: styurin

fbshipit-source-id: b69fb75