Allow MacroArg to take advantage of query support for precomputed cache
Summary:
Precomputed macro work must be stored by the caller since the expanders are stateless.
This also helps the cache to be tied to the lifetime of the rule rather than the lifetime of the macro expander.
This change opts-in the MacroArg to the caching system.

Test Plan:
Testing of the precomputation directly is in the diff below.
Profiling this change shows a several second win perf win for a medium java build. build_helper output coming shortly.

Reviewed By: styurin

fbshipit-source-id: 912f699