Refactor to decrease the size of the parameter lists
Summary:
I disapprove of the monster-parameter-list pattern. It makes for big code that's hard to follow and tedious to maintain; passing a single `interface` that can 'reach into' calling code to extract state on-demand is slightly cheaper, and makes it a lot easier to update the parameter list, especially when (as here) the parameters are passed through (and stored in) a 'helper' class.

This diff introduces the smaller, more-maintainable parameter-extraction pattern. The interfaces are Just Good Enough for a 1st pass and can undoubtedly be improved and/or generalized, but the code is already smaller, clearer, and slightly faster. This diff doesn't touch iOS project generation, and makes only minimal changes to the current standard `buck project` IJ generator.

Test Plan: `buck test //test/com/facebook/buck/ide/intellij/projectview && buck test //test/com/facebook/buck/ide/intellij`

Reviewed By: styurin

fbshipit-source-id: f131c65