Log failure to read RECORDED_PATHS
Summary:
There have been reports of failure to read RECORDED_PATHS, but I
haven't seen it in any logs.  It might be getting rolled up into a generic
"NoSuchFileException", though, so I want to force some specific logging for this
case to make it easier to spot and debug.

Test Plan: Delete a RECORDED_PATHS from my buck-out and try a build

Reviewed By: asp2insp

fbshipit-source-id: 9772ffa