Consolidate the TreePathScanner and ElementScanner in TreeBackedEnter
Summary:
We're going to need `TreePath`s in the `TreeBackedElement`s after all, so that we
can look up `javac`'s opinion of the types of various subexpressions. This commit
is just smashing a couple classes together in preparation for a change to
construct the `TreePath`s in an efficient manner.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 3be7979