ProjectFilesystem.mkdirs should allow path to be a symlink pointing to a directory
Summary:
Previously, if the path is a symlink, it will throw a
FileAlreadyExistsException, even if the symlink points to a valid directory.
This has the effect of disallowing buck-out to be a symlink.

This diff removes that limitation by testing if the path is a directory symlink
if the exception is thrown.

This alone doesn't solve all the problems with buck-out being a symlink, since
there's still cases (namely c++ header maps), where we try to `..` our way out
of buck-out, which doesn't work as expected if buck-out is a symlink.

Test Plan: Added a test.

Reviewed By: Coneko

fbshipit-source-id: e2d06af