Revert DirectHeaderMaps to generate absolute paths
Summary:
This was originally made to generate relative paths because
HeaderPathNormalizer can't deal with absolute paths. The limitation is lifted,
so this can once again use absolute paths. The previous behavior of being
relative to buck-out makes it impossible for buck-out to be a symlink, as ..
from such a buck-out can be anywhere.

Test Plan: CI

Reviewed By: dinhviethoa

fbshipit-source-id: 8770ae8