A merged view over SortedSets
Summary:
There are several places where we proactively merge and memoize SortedSets
(such as BuildRuleParams.totalBuildDeps).

We also have DependencyAggregation, a BuildRule with really weird
semantics, to avoid the memory overhead of copying many
similar-but-not-quite-the-same SortedSets.

This view will allow us to simplify these cases, and claim back some
memory as we do.

Test Plan: Added tests

Reviewed By: Coneko

fbshipit-source-id: 76d4a0f