Run relinker rules globally on all so files and then copy based on modulesinstead of running relinker on a per-module basis
Summary:
the relinker currently is scoped to only look at linkables within a single app module, however, linkables within an app module can depend on linkables in other app modules, including the root.
This change keeps track of the apk modules the linkable belong to, but still precess re-linking for all linkables in the app.  After the relinker is complete, we copy the linkables based on their app modules to the proper location

Test Plan: added test cases for modular re-linked linkables

fbshipit-source-id: e02cd73