Add forEachThrowing and forEachOrderedThrowing to RichStream
Summary:
For sink operations, we can safely throw the operation's exception, since the
operation is performed immediately. This diff adds throwing variants of
`forEach` and `forEachOrdered`.

The implementation wraps the consumer inside another which wraps any checked
exceptions that are thrown. The wrapped exception is unwrapped and re-thrown
outside the delegated call. Care is taken to ensure that only the expected
exception wrapper is unwrapped.

Test Plan: Unit Tests.

Reviewed By: coneko

fbshipit-source-id: fbe13d5