Add --show-transitive-rulekeys to buck targets
Summary:
This only takes effect when --show-rulekeys is specified.  It will print
the transitive deps of the specified targets and their rule keys.

The intention is to use this to compose a list of artifacts to pre-fetch
from cache.

Test Plan: Ran it on an android_binary.  Saw dexes and abis.

Reviewed By: marcinkosiba

fbshipit-source-id: c0024d1