Fix R.txt rewriting of styleables
Summary:
Android expects that the styleable arrays in R.java are in sorted order.
ExoResourcesRewriter.rewriteRDotTxt() needs to make sure that the
rewritten arrays are sorted. It then needs to update the indices that
follow (there is a named value mapping to each entry in the array).

Test Plan: CI, added test

Reviewed By: dreiss

fbshipit-source-id: 2643aab