Improve assumption Precondition checking
Summary:
This code is written assuming that certain things, that are technically
allowed by the specification (if it can be called that), don't actually
occur in practice. This just adds verification of a couple of those
assumptions so that we get a hard crash if they are violated rather
than possibly silently doing the wrong thing.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 68d71a8