Retrying artifact cache
Summary:
The previous behavior of fetching a remote artifact is that:
- it tries to connect
- when it fails, it retries
- then, it receives the data
- if it fails receiving the data, it fails and it doesn't retry

This diff will retry fetching a remote artifact if even if receiving the data fails.

Test Plan:
```
$ buck build app
```

before:
on a midly broken connection: sometimes, it failed receiving the data

after:
on a midly broken connection: sometimes, it retries receiving the data, then, it never fails fetching artifacts.

Reviewed By: yiding

fbshipit-source-id: 36bf99d