Ability to change the version using an environment variable with buck.pex
Summary:
When buck is built with buck, the output is in buck-out/gen/programs/buck.pex
In that situation we can't use a fake buck version when using that binary.

This change will allow us to set an environment variable FAKE_BUCK_VERSION to simulate a given buck version.

Test Plan:
```
$ FAKE_BUCK_VERSION=[buck-version-of-a-repo] ~/BuckRepo/buck/buck-out/gen/programs/buck.pex build app
```

Make sure cache matches.

Reviewed By: bertmaher

fbshipit-source-id: fffcace