Add support for kotlin and java mixed sources.
Summary:
This adds a javac step to the KotlincToJarStepFactory.
If any java files are found it will invoke javac, otherwise it won't do
anything.

This also adds java annotations processors to kotlin library rules.

This also fixes an issue where the bootclasspath was not being set properly for kotlin rules.

For now, this only works if the java annotation processors do not reference
kotlin code and the kotlin code does not rely on annotation processors.

This fixes #1037

This probably isn't ready to merge yet until a stable version of kotlinc has support for passing java files directly to it. Currently you can use `1.1.3-eap-68` or newer to test this out.
Closes https://github.com/facebook/buck/pull/1358

Reviewed By: dsyang

Pulled By: dsyang

fbshipit-source-id: c31fb5c