Limit the number of targets aggregated to module with Android resources
Summary:
Android plugin tries to save a list of resource folders in a UTF8 string using DataOutputStream and it has a limitation of 64k. As a result, when we aggregate all modules into one modules, the plugin fails to save because the list is too big.

Adding a new config option to limit the size of the module.

Test Plan:
Unit tests and local testing.

Run IntelliJ with a project with resource aggregation and without limit, see that IntelliJ removes Android facet from the iml file.

Run IntelliJ with a project with resource aggregation and with limit, see that all resources are resolved.

Reviewed By: asp2insp

fbshipit-source-id: 5b85466