Don't scrub the padding of the symbol name table.
Summary:
The padding os the symbol name table is not uninitialised memory
since approximately Xcode 8 so it doesn't need to be scrubbed
anymore.

Unfortunately we still can't just use `ZERO_AR_DATE` since we still need to scrub uid and gid, but that code is shared with the GNU archiver so the cost is split.

Test Plan:
CI
AppleLibraryIntegrationTest.testAppleLibraryIsHermetic

Reviewed By: k21

fbshipit-source-id: 4a3221d