Record and publish the number of files materialized locally and remotely
Summary:
Stampede build slaves now keep track of the following stats:
- number of files materialized from local cache
- number of files materialized remotely
- total time spent materializing files remotely

These stats are published to the event bus at the end of the build. Additionally, the number of files materialized (across all build slaves) is displayed on the SuperConsole of the Buck client that kicked off the distributed build.

Test Plan: new unit tests

Reviewed By: marcinkosiba

fbshipit-source-id: aac008e