Work around bug in how NuProcess closes stdin
Summary:
On OSX, `BasePosixProcess.closeStdin(true)` tries to add the process to
a bounded close queue before actually closing the file descriptor.  However, it
appears that the only way that queue is ever drained is if a write is requested.
So, the queue fills up and starts throwing before it can close the fd... but
since we closeStdin in an onStart, the exception gets caught and ignored inside
nuprocess itself.

This shows up in a pretty interesting manner: after you've opened 512 * ncpus /
2 subprocesses (that's the size of the queue times the number of queues), you
start leaking pipes.

`closeStdin(false)` doesn't have this problem, so use that instead.  (Also, we
can't rely on the implicit close that nuprocess does during `onExit`, because it
uses `closeStdin(true)`.

Test Plan:
This no longer shows unbounded pipe growth:
```
for i in `seq 52`; do
    buck server status --http-port
    lsof -p $(pgrep -f buckd) | grep PIPE | wc -l
done
```

Reviewed By: yiding

fbshipit-source-id: 7e0d2f2