merge buck doctor & rage commands
Summary:
This merges the code of buck rage and buck doctor commands. The commands used to
do very similar things and duplication of many of the classes wasn't fun. The new behaviour
is altered a bit but nothing really important. Now rage links to doctor and doctor is the only
command that is actually executed.

I kept the model where you can `rage/doctor` only one command. Also some configuration
names changed to make it easier to setup the different report, analysis endpoints.

I added the necessary documentation that was missing from rage so that should help.

Test Plan:
- Necessary tweaks to the existing tests.
- Manual testing.

fbshipit-source-id: 287d133