Utils manipulate BuildTargets not BuildRuleParams
Summary:
I'm about to start referring directly to the BuildTargets, rather than
the BuildRuleParams, so I need to not require a BuildRuleParams to use
these utilities.a

Note that there is a lot of silly removing and re-appending of
flavors, where the underlying BuildRuleParams could be kept and
referenced, rather than pretending that they are actually mutated with
each re-assignment. I will make some of these go away soon, but don't
want to get too distracted by them.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 821cd6c